backend/requirements.txt
fastapi
uvicorn
supabase
pandas
numpy
scikit-learn
python-dotenv
openpyxl
python-multipart